+++ packages/disk/src/sandbox-process.ts
@@ -147,6 +147,7 @@ export class SandboxProcesses {
     options: SandboxProcessStartOptions = {},
   ): Promise<SandboxProcess> {
     const process = new SandboxProcess(
+      this._sandboxId,
       "",
       0,
       options.onOutput,
@@ -177,6 +178,7 @@ export class SandboxProcesses {
   ): Promise<SandboxProcess> {
     const offset = options.offset ?? 0;
     const process = new SandboxProcess(
+      this._sandboxId,
       processId,
       offset,
       options.onOutput,
@@ -263,6 +265,7 @@ export class SandboxProcess {
 
   /** @internal */
   constructor(
+    readonly sandboxId: string,
     processId: string,
     cursor: number,
     onOutput: SandboxProcessOutputHandler | undefined,